+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update to circle-stdlib v15.9/Circle Step 44.2.
 - Update ARM toolchains to 10.3-2021.07.
 - Update to libmt32emu v2.5.3.
-- Update to FluidSynth v2.2.3.
+- Update to FluidSynth v2.2.4.
 
 ## [0.10.3] - 2021-09-03